basic cut down
