+++ apps/code-examples/src/app/code-examples/datetime/date-range-picker/basic/demo.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnDestroy, OnInit, inject } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import {
   AbstractControl,
   FormBuilder,
@@ -7,18 +7,8 @@ import {
   FormGroup,
   FormsModule,
   ReactiveFormsModule,
-  ValidationErrors,
 } from '@angular/forms';
-import {
-  SkyDateRangeCalculation,
-  SkyDateRangeCalculatorId,
-  SkyDateRangeCalculatorType,
-  SkyDateRangePickerModule,
-  SkyDateRangeService,
-} from '@skyux/datetime';
-
-import { Subject } from 'rxjs';
-import { distinctUntilChanged, takeUntil } from 'rxjs/operators';
+import { SkyDateRangePickerModule } from '@skyux/datetime';
 
 @Component({
   standalone: true,
@@ -31,139 +21,22 @@ import { distinctUntilChanged, takeUntil } from 'rxjs/operators';
     SkyDateRangePickerModule,
   ],
 })
-export class DemoComponent implements OnInit, OnDestroy {
+export class DemoComponent {
   protected get reactiveRange(): AbstractControl | null {
     return this.formGroup.get('lastDonation');
   }
 
-  protected calculatorIds: SkyDateRangeCalculatorId[] | undefined;
   protected dateFormat: string | undefined;
+  protected disabled = false;
   protected formGroup: FormGroup;
-
-  #ngUnsubscribe = new Subject<void>();
-
-  readonly #dateRangeSvc = inject(SkyDateRangeService);
+  protected hintText =
+    'Donations received today are updated at the top of each hour.';
+  protected labelText = 'Last donation';
+  protected required = true;
 
   constructor() {
     this.formGroup = inject(FormBuilder).group({
-      lastDonation: new FormControl(),
+      lastDonation: new FormControl({ value: '', disabled: this.disabled }),
     });
-  }
-
-  public ngOnInit(): void {
-    // Watch for status changes.
-    this.reactiveRange?.statusChanges
-      .pipe(distinctUntilChanged(), takeUntil(this.#ngUnsubscribe))
-      .subscribe((status) => {
-        console.log(
-          'Date range status change:',
-          status,
-          this.reactiveRange?.errors,
-        );
-      });
-
-    // Watch for value changes.
-    this.reactiveRange?.valueChanges
-      .pipe(distinctUntilChanged(), takeUntil(this.#ngUnsubscribe))
-      .subscribe((value: SkyDateRangeCalculation) => {
-        console.log('Date range value change:', value);
-      });
-  }
-
-  public ngOnDestroy(): void {
-    this.#ngUnsubscribe.next();
-    this.#ngUnsubscribe.complete();
-  }
-
-  protected toggleDisabled(): void {
-    if (this.formGroup.disabled) {
-      this.formGroup.enable();
-    } else {
-      this.formGroup.disable();
-    }
-  }
-
-  protected resetForm(): void {
-    this.dateFormat = undefined;
-    this.calculatorIds = undefined;
-    this.formGroup.reset();
-    this.formGroup.markAsPristine();
-    this.formGroup.markAsUntouched();
-  }
-
-  protected setRange(): void {
-    const range: SkyDateRangeCalculation = {
-      calculatorId: SkyDateRangeCalculatorId.SpecificRange,
-      startDate: new Date('1/1/2012'),
-      endDate: new Date('1/1/2013'),
-    };
-
-    this.reactiveRange?.setValue(range);
-  }
-
-  protected setInvalidRange(): void {
-    const range: SkyDateRangeCalculation = {
-      calculatorId: SkyDateRangeCalculatorId.SpecificRange,
-      startDate: new Date('1/1/2013'),
-      endDate: new Date('1/1/2012'),
-    };
-
-    this.reactiveRange?.setValue(range);
-    this.reactiveRange?.markAsTouched();
-  }
-
-  protected setInvalidDates(): void {
-    const range: SkyDateRangeCalculation = {
-      calculatorId: SkyDateRangeCalculatorId.SpecificRange,
-      startDate: 'invalid' as never as Date,
-      endDate: 'invalid' as never as Date,
-    };
-
-    this.reactiveRange?.setValue(range);
-  }
-
-  protected setCalculatorIds(): void {
-    const calculator = this.#dateRangeSvc.createCalculator({
-      shortDescription: 'Since 1999',
-      type: SkyDateRangeCalculatorType.Relative,
-      getValue: () => {
-        return {
-          startDate: new Date('1/1/1999'),
-          endDate: new Date(),
-        };
-      },
-    });
-
-    this.calculatorIds = [
-      calculator.calculatorId,
-      SkyDateRangeCalculatorId.SpecificRange,
-      SkyDateRangeCalculatorId.LastFiscalYear,
-    ];
-  }
-
-  protected setCustomError(): void {
-    const calculator = this.#dateRangeSvc.createCalculator({
-      shortDescription: 'Date before today',
-      type: SkyDateRangeCalculatorType.Before,
-      validate: (value): ValidationErrors | null => {
-        if (value && value.endDate && value.endDate > new Date()) {
-          return {
-            dateIsAfterToday: true,
-          };
-        }
-        return null;
-      },
-      getValue: () => {
-        return {
-          endDate: new Date('1/2/2050'),
-        };
-      },
-    });
-
-    this.calculatorIds = [calculator.calculatorId];
-  }
-
-  protected setDateFormat(): void {
-    this.dateFormat = 'YYYY-MM-DD';
   }
 }
