+++ apps/code-examples/src/app/home/home.component.html
@@ -123,6 +123,11 @@
         Date range picker
         <ul>
           <li><a routerLink="datetime/date-range-picker/basic">Basic</a></li>
+          <li>
+            <a routerLink="datetime/date-range-picker/custom-calculator"
+              >Custom calculator</a
+            >
+          </li>
         </ul>
       </li>
       <li>
